web/backend/user-service/requirements.txt
aiomysql
bcrypt
tornado
pyjwt